Crate uorustlibs

Modules

anim
Methods for reading animated characters out of anim.mul/anim.idx

art
Methods for reading tile and static data out of art.mul

color

gump

hues
Methods for reading hue data out of hues.mul

map

mul_reader
Methods for reading from standardized Mul and Idx files

skills
Skill objects represent named skills that appear in UO's Skills menu. They also contain a flag denoting whether they are clicked to activate

texmaps
Methods for reading texture data out of texmaps.mul

tiledata

utils