Module uorustlibs::mul_reader

Methods for reading from standardized Mul and Idx files

IDX files are defined as |index:u32|size:u32|opt1:u16|opt2:u16|

Where index and size represent references into the equivalent Mul file

Index values of 0xFEFEFEFF are considered undefined, and should be skipped

Structs

MulReader
Read Mul records out of am idx and a mul

MulRecord
An individual record, read from a Mul file

MulWriter
Write new records onto existing Mul and Idx files

Enums

MulWriterMode