Struct uorustlibs::mul_reader::MulRecord

pub struct MulRecord {
    pub data: Vec<u8>,
    pub start: u32,
    pub length: u32,
    pub opt1: u16,
    pub opt2: u16,
}

An individual record, read from a Mul file

Fields
data: Vec<u8>

Raw Mul data

start: u32

The index position in the Mul of this item

length: u32

The total length in the Mul of this item

opt1: u16

An implementation-specific variable

opt2: u16

An implementation-specific variable